Split out UsaNeko/Peace into common abstract base class and two implementations.
